Create LWPostsPreviewTooltip
